back to ps1, attempting to globalstar

+++ smpnsc.py
@@ -4945,13 +4945,15 @@ class smp:
         #    print f, m
         print '-'*100
         #sys.exit()
+
         if self.savezptstamps:
             print 'star fit stamps saved in ', self.zptstamps
             pdf_pagesc.close()
-            self.tmpwriter.cp('starfits_' + str(thismjd) + '.pdf',
-                              '/'.join(longimfile.split('/')[:-1]) + '/starfitstamps.pdf')
-            self.tmpwriter.cp('starfits_' + str(thismjd) + '.pdf',self.zptstamps)
-            os.popen('rm starfits_' + str(thismjd) + '.pdf')
+            if not self.snparams.survey == 'PS1':
+                self.tmpwriter.cp('starfits_' + str(thismjd) + '.pdf',
+                                  '/'.join(longimfile.split('/')[:-1]) + '/starfitstamps.pdf')
+                self.tmpwriter.cp('starfits_' + str(thismjd) + '.pdf',self.zptstamps)
+                os.popen('rm starfits_' + str(thismjd) + '.pdf')
             #raw_input()
             if self.fermilog:
                 self.tmpwriter.appendfile('saved starfit stamps to pdf\n', self.fermilogfile)
